opensearch/requirements.txt
# Python 3.8.10 is used in the project
opensearch-py==2.7.1
opensearch-py-ml==1.1.0
pandas==1.5.3
pyarrow==17.0.0
ipython==7.25.0
deprecated==1.2.14
sentence-transformers==3.1.1
mkdocs==1.6.1
mkdocs-material==9.5.37
openai==1.49.0
python-dotenv==1.0.1
torch==2.4.1
pyyaml==6.0.2
accelerate==0.34.2
sentence_transformers==3.1.1
onnx==1.16.2

# for pdf print
WeasyPrint==61.2
pydyf==0.8.0 # higher versions issues - https://github.com/Kozea/WeasyPrint/issues/2200
pdfkit==1.0.0

# open telemetry
Flask==3.0.3
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp==1.27.0

# Faker
Faker==30.8.1
